requirements.txt
numpy==1.26.4
pinecone==5.0.1
PyPDF2==2.10.5
python-dotenv==0.21.0
Requests==2.32.3
torch==2.3.0
transformers==4.44.1
fastapi==0.103.0
mangum==0.17.0
uvicorn==0.30.6
requests==2.32.3